+++ src/core/electric_utility.jl
@@ -33,6 +33,7 @@
 Base.@kwdef struct ElectricUtility
     outage_start_timestep::Int=0  # for modeling a single outage, with critical load spliced into the baseline load ...
     outage_end_timestep::Int=0  # ... utiltity production_factor = 0 during the outage
+    allow_simultaneous_export_import::Bool=true  # if true the site has two meters (in effect)
     # variables below used for minimax the expected outage cost,
     # with max taken over outage start time, expectation taken over outage duration
     outage_start_timesteps::Array{Int,1}=Int[]  # we minimize the maximum outage cost over outage start times
@@ -46,6 +47,7 @@ end
 Base.@kwdef struct ElectricUtility
     outage_start_timestep::Int=0  # for modeling a single outage, with critical load spliced into the baseline load ...
     outage_end_timestep::Int=0  # ... utiltity production_factor = 0 during the outage
+    allow_simultaneous_export_import::Bool=true  # if true the site has two meters (in effect)
     # variables below used for minimax the expected outage cost,
     # with max taken over outage start time, expectation taken over outage duration
     outage_start_timesteps::Array{Int,1}=Int[]  # we minimize the maximum outage cost over outage start times